Move page components out of SeversPage
